refactor(juntos-bot): clear files and make a separation of concerns
